santad: Move event providers into a new group, make ES connection logic smarter
